third_party/WebKit/LayoutTests/virtual/outofblink-cors/external/wpt/service-workers/service-worker/clients-matchall-client-types.https-expected.txt

This is a testharness.js-based test.
FAIL Verify matchAll() with window client type assert_unreached: unexpected rejection: assert_array_equals: property 2, expected "https://web-platform.test:8444/service-workers/service-worker/resources/clients-matchall-client-types-iframe.html" but got "https://web-platform.test:8444/service-workers/service-worker/resources/url-modified-via-pushstate.html" Reached unreachable code
FAIL Verify matchAll() with {window, sharedworker, worker} client types promise_test: Unhandled rejection with value: object "Error: wait_for_state must be passed a ServiceWorker"
Harness: the test ran to completion.
